refactoring: create a generic CrossingGenerator for RHW/NMW/etc.
